% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_gp_warp.R
\name{gp_template_warp}
\alias{gp_template_warp}
\title{Template Gaussian Process Warping}
\usage{
gp_template_warp(y_list, feat_list, template_feats, niter = 5000,
  nburn = 10000, int_p = 20, asig = 0.1, bsig = 0.1, at = 0.1,
  bt = 0.1, al = 0.1, bl = 0.1, aa = 0.01, ba = 0.01,
  progress = TRUE)
}
\arguments{
\item{y_list}{list of curves with equal length}

\item{feat_list}{list of feature vectors with equal length}

\item{template_feats}{a feature vector with length mathing those in feat_list}

\item{niter, nburn}{MCMC iterations and burns. Total run time is the sum.}

\item{int_p}{number of internal knots specified for the data curves}

\item{asig, bsig, at, bt, al, bl, aa, ba}{hyperparameters for inverse gamma priors on model variances}

\item{progress}{logical; indicates whether or not to print percent progress.}
}
\value{
a list of warped curves, the estimated mean curve, and a vector of MCMC
  acceptance rates. Curve list elements are vectors with length matching
  the elements of y_list.
}
\description{
Performs hierarchical curve warping by modeling the warping functions with GP.
}
