% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{interp_spline}
\alias{interp_spline}
\title{Evenly Spaced Curve Interpolation using Splines}
\usage{
interp_spline(y, x = NULL, nout = NULL)
}
\arguments{
\item{y}{numeric vector giving curve values}

\item{x}{numeric vector giving input times. Defaults to evenly spaced values}

\item{nout}{number of points defining resultant interpolated curve. Defaults to length of y}
}
\description{
Interpolates curve values using cubic interpolation splines.
The resulting evaluations correspond to an evenly spaced grid of input points.
Often used to time-normalize registered curves onto a common time domain.
Note the quality of interpolation decreases with inadequate density of defining points.
If ties are found among input times, linear interpolation is used instead of splines.
}
