% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{K1}
\alias{K1}
\title{First Order Penalty Matrix}
\usage{
K1(dim)
}
\arguments{
\item{dim}{dimension of output matrix. Must be at least three.}
}
\description{
Generates a first order penalty matrix discussed by Lang and Brezger (2012).
In short, when used as a prior precision matrix for B-splines, this matrix imposes
a roughness penalty. Note that this matrix is not full rank.
}
