% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{K2}
\alias{K2}
\title{Second Order Penalty Matrix}
\usage{
K2(dim)
}
\arguments{
\item{dim}{dimension of output matrix. Must be at least four.}
}
\description{
Creates a second order penalty matrix used in fitting penalized b-splines.
Courtesy of CITATION NEEDED. Note that this function produces singular matrices.
}
