% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_warp.R
\name{landmark_warp}
\alias{landmark_warp}
\title{Landmark Warping}
\usage{
landmark_warp(y_list, feat_list, template_feats)
}
\arguments{
\item{y_list}{list of curves with equal length}

\item{feat_list}{list of feature vectors with equal length}

\item{template_feats}{a feature vector with length mathing those in feat_list}
}
\value{
a list of warped curves and a pointwise mean curve. List elements are
  vectors with length matching the elements of y_list.
}
\description{
Warps a list of curves by matching landmarks to provided template landmarks.
}
