% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{register}
\alias{register}
\title{Register Curves Given Warping Functions}
\usage{
register(y, wtime)
}
\arguments{
\item{y}{numeric matrix; columnns should correspond to observed time normalized curves}

\item{wtime}{numeric matrix or three dimensional array; columns should correspond to warping function realizations and MCMC iterates should correspond to the third array dimension (if applicable).}
}
\description{
Computes registered curves from either given time warping vectors or an
MCMC array of time warping vectors though the inverse warping function.
}
