% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telesca_warp.R
\name{telesca_warp}
\alias{telesca_warp}
\title{Telesca Hierarchical Warping}
\usage{
telesca_warp(y_list, niter = 5000, nburn = 10000, int_q = 5,
  int_p = 20, asig = 0.1, bsig = 0.1, at = 0.1, bt = 0.1,
  al = 0.1, bl = 0.1, progress = TRUE, debug = FALSE)
}
\arguments{
\item{y_list}{list of curves with equal length}

\item{niter, nburn}{MCMC iterations and burns. Total run time is the sum.}

\item{int_q}{number of internal knots specified for the warping function}

\item{int_p}{number of internal knots specified for the data curves}

\item{asig, bsig, at, bt, al, bl}{hyperparameters for inverse gamma priors on model variances}

\item{progress}{logical; indicates whether or not to print percent progress.}

\item{debug}{logical; stops function right before return and enters debug mode.}
}
\value{
a list of warped curves, the estimated mean curve, and a vector of MCMC
  acceptance rates. Curve list elements are vectors with length matching
  the elements of y_list.
}
\description{
Implements a method of bayesian hierarchical curve registration described in a paper
by Telesca (date).
}
