% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_gp_warp.R
\name{template_gp_warp}
\alias{template_gp_warp}
\title{Template Gaussian Process Warping}
\usage{
template_gp_warp(y_list, feat_list, template_feats, niter = 1000,
  nburn = 1000, int_p = 20, asig = 0.1, bsig = 0.1, at = 0.1,
  bt = 0.1, al = 0.1, bl = 0.1, aa = 1, ba = 30, be = 2,
  tune = 2, progress = TRUE, debug = FALSE)
}
\arguments{
\item{y_list}{list of curves with equal length}

\item{feat_list}{list of feature vectors with equal length}

\item{template_feats}{a feature vector with length mathing those in feat_list}

\item{niter, nburn}{MCMC iterations and burns. Total run time is the sum.}

\item{int_p}{number of internal knots specified for the data curves}

\item{asig, bsig, at, bt, al, bl}{hyperparameters for inverse gamma priors on model variances}

\item{aa, ba}{hyperparameters for the uniform prior on inverse range parameter}

\item{tune}{tuning parameter for the proposal normal distribution of alpha.}

\item{progress}{logical; indicates whether or not to print percent progress.}

\item{debug}{logical; stops function right before return and enters debug mode.}
}
\value{
a list of warped curves, the estimated mean curve, and a vector of MCMC
  acceptance rates. Curve list elements are vectors with length matching
  the elements of y_list.
}
\description{
Performs hierarchical curve warping by modeling the warping functions with GP.
}
