% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp.R
\name{warp_bhcr}
\alias{warp_bhcr}
\title{Bayesian Hierarchical Curve Registration}
\usage{
warp_bhcr(y, p = 15, q = 5, n_iter = 15000, n_burn = 50000,
  n_thin = 1, a_eps = 0.1, b_eps = 0.1, a_a = 0.1, b_a = 0.1,
  a_c = 0.1, b_c = 0.1, a_tau = 0.1, b_tau = 0.1, a_lam = 0.1,
  b_lam = 0.1)
}
\description{
Bayesian Hierarchical Curve Registration
}
