% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp.R
\name{warp_template}
\alias{warp_template}
\title{Template Prior Curve Registration}
\usage{
warp_template(
  y,
  lmk_inds,
  ref_lmk_inds,
  p = 10,
  n_iter = 15000,
  n_burn = 10000,
  n_thin = 1,
  a_eps = 0.1,
  b_eps = 0.1,
  a_a = 0.1,
  b_a = 0.1,
  a_c = 0.1,
  b_c = 0.1,
  a_tau = 0.1,
  b_tau = 0.1,
  a_lam = 0.1,
  b_lam = 0.1
)
}
\description{
Template Prior Curve Registration
}
