% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{interp_spline}
\alias{interp_spline}
\title{Spline Curve Interpolation}
\usage{
interp_spline(x, y, nout = length(y))
}
\arguments{
\item{x, y}{coordinate vectors on which to fit a cubic spline.}

\item{nout}{number of points to interpolate out}
}
\description{
Interpolates out equally spaced values by fitting a cubic spline on given points and
matching the endpoints. Often used to redefine a curve with warped time axis on an
even time basis. Note that quality of interpolation decreases with inadequate density
of defining points. If ties are found, linear interpolation is used.
}
