% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{K1}
\alias{K1}
\title{First Order Penalty Matrix}
\usage{
K1(dim)
}
\arguments{
\item{dim}{dimension of output matrix. Must be at least three.}
}
\description{
Creates a first order penalty matrix used in fitting penalized b-splines.
Courtesy of CITATION NEEDED. Note that this function produces singular matrices.
}
