% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_warp.R
\name{landmark_warp}
\alias{landmark_warp}
\title{Landmark Warping}
\usage{
landmark_warp(y_list, feat_list, template_feats, niter = 1000,
  nburn = 1000, int_p = 20, asig = 0.1, bsig = 0.1, at = 0.1,
  bt = 0.1, progress = TRUE, debug = FALSE)
}
\arguments{
\item{y_list}{list of curves with equal length}

\item{feat_list}{list of feature vectors with equal length}

\item{template_feats}{a feature vector with length mathing those in feat_list}

\item{niter, nburn}{MCMC iterations and burns. Total run time is the sum.}

\item{int_p}{number of internal knots specified for the data curves}

\item{asig, bsig, at, bt}{hyperparameters for inverse gamma priors on model variances}

\item{progress}{logical; indicates whether or not to print percent progress.}

\item{debug}{logical; stops function right before return and enters debug mode.}
}
\value{
a list of warped curves and an estimated mean curve. List elements are
  vectors with length matching the elements of y_list.
}
\description{
Warps a list of curves by matching landmarks to provided template landmarks.
}
